Add files via upload
In this lab, I got hands-on practice with the logistic regression model, a popular and effective classification model. Understanding logistic regression and being able to apply it to classification tasks is essential because logistic regression models form the fundamentals of neural networks.

We will use a real-world dataset that contains detailed nutrition information about food items for people with diabetes. The objective is to classify whether a diabetic patient should choose More Often, Less Often, or In Moderation for a specific food item based on the nutrition information in the dataset.

Objectives
Preprocess and generate training and testing datasets
Train and fine-tune logistic regression models
Interpret trained logistic regression models
Evaluate trained logistic regression models